web/requirements.txt
Flask
flask-mysqldb
Werkzeug
mysql-connector-python